# FilmApi/DTOs/Movie/CreateMovieDto.cs
using FilmApi.Models;
using System.ComponentModel.DataAnnotations;

namespace FilmApi.DTOs.Movie
{
    public class CreateMovieDto
    {
        public string Title { get; set; }

        public string Genre { get; set; }
        public int ReleaseYear { get; set; }
        public string Director { get; set; }

        public string Picture { get; set; }

        public string Trailer { get; set; }

    }
}